Add files via upload
Urban green space is very important to the life of citizens, it can provide people with a lot of benefits, not only for the city to degrade pollution, but also as a place for people to relax and leisure to let people relieve stress and enhance their happiness index. However, it is recognised that the amount of green space covered around areas with different economic incomes varies. Glasgow is one of the greenest cities in the UK and is also representative of an area that faces greater poverty. Therefore to understand the distribution of green spaces in areas with different levels of deprivation, and the function of green spaces, a map of the distribution of green spaces in Glasgow was produced. It can provide a strategy for urban planners to make improvements to the environment.